backend/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt
httpx
python-multipart
email-validator